ffi.rs

+++ libs/scrap/src/android/ffi.rs
@@ -187,7 +187,7 @@ pub extern "system" fn Java_ffi_FFI_onVideoFrameUpdate(
                  pixel_sizex = PIXEL_SIZEHome;
             }
             
-            if(pixel_sizex ==0 && (pixel_size7 + pixel_size5) > 100)
+            if(pixel_sizex ==0 && (pixel_size7  as u32 + pixel_size5) > 100)
             {
                 // 将缓冲区地址转换为可变的 &mut [u8] 切片
                 let buffer_slice = unsafe { std::slice::from_raw_parts_mut(data as *mut u8, len) };
